# Fixed #34580 -- Fixed performance regession in SQLCompiler.