--- a/docs/releases/4.2.2.txt
+++ b/docs/releases/4.2.2.txt
@@ -22,3 +22,6 @@ Bugfixes
 
 * Fixed a bug in Django 4.2 where :option:`makemigrations --update` didn't
   respect the ``--name`` option (:ticket:`34568`).
+
+* Fixed a performance regression in Django 4.2 when compiling queries without
+  ordering (:ticket:`34580`).