This is @charettes's commit not mine, so commit author will need changing if this works. See [Ticket #34580](https://code.djangoproject.com/ticket/34580)

I'm just opening this to see if the benchmark script can see a performance gain (it's about a 20%-30% improvement locally). 


```
       before           after         ratio
     [ca5d3c99]       [2aa6996a]
     <form_docs>       <SQLCompiler>
-     1.81±0.01ms         1.43±0ms     0.79  query_benchmarks.query_exists.benchmark.QueryExists.time_query_exists
-         876±7μs          625±8μs     0.71  query_benchmarks.query_aggregate.benchmark.QueryAggr.time_aggregate
-     1.30±0.01ms          914±9μs     0.70  query_benchmarks.query_count.benchmark.QueryCount.time_query_count

```