@smithdc1 2023-05-21
No problem -- I'm not sure it's been used all that much so this was an interesting case to see if it is useful in practice. 

It seems that running the workflow on github's infrastructure led to no significant results being found. However, running locally again I see some quite large benefits from this.  

```
       before           after         ratio
     [ca5d3c99]       [2aa6996a]
     <form_docs>       <SQLCompiler>
-     1.83±0.01ms      1.41±0.01ms     0.77  query_benchmarks.query_exists.benchmark.QueryExists.time_query_exists
-         874±2μs          621±8μs     0.71  query_benchmarks.query_aggregate.benchmark.QueryAggr.time_aggregate
-     1.35±0.02ms          905±4μs     0.67  query_benchmarks.query_count.benchmark.QueryCount.time_query_count
```