@charettes 2023-05-21
It seems like a valuable change so I'll accept the ticket but I'm not surprised that we get weird results on Github CI; any form of virtualized / containerized process is doomed to provide inaccurate results.